requirements.txt
anthropic>=0.40.0
requests>=2.32.0
beautifulsoup4>=4.12.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
feedparser>=6.0.11
reportlab>=4.2.0
matplotlib>=3.9.0
Pillow>=10.4.0
python-dotenv>=1.0.1
pytz>=2024.1
typing-extensions>=4.12.0
squarify>=0.4.3

# Newsletter / API (Fase 1)
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
Flask>=3.0.0
Werkzeug>=3.0.0
sendgrid>=6.11.0

# Auth / Pagos / Alertas (Fase 2)
flask-jwt-extended>=4.6.0
bcrypt>=4.1.0
stripe>=9.0.0
apscheduler>=3.10.0

# Producción — Railway proporciona SSL/HTTPS automáticamente en dominios .railway.app
gunicorn>=22.0.0
